Add files and their content
The project had no content, so files with some basic structure are
required.

As the project was lacking in content add an index, img folder, recipes folder,
and a first recipe with some content in the index file and the recipe file.

The the folders, files and content added resulted in a basic but good structured
website.