Create new recipe pages with their content and links to reach them
Content needs to be added to the other recipe pages and links at the index
need to be created in order to reach them.

As the website needed more recipes, create new pages, add content to them and create
more links at the index to reach them.

As a result, the website has 2 new pages for 2 new recipes that can be reached and their
content is already there ready to be displayed.